MotorModule finished first pass. Start integrate MotorModule in Main
The first pass of MotorModule is done. This state is saved in the MotorModule.ino file.
The files MotorModule.cpp and MotorModule.h are needed for the inclusion to the SmartWheels.ino.
Currently the project is buildable but the Test functions don't run properly on the hardware.
No reaction is seen on the Hardware. Debug message this far point to the conclusion that the code is executed.

TODO: Fix the integration to the SmartWheels.ino file
TODO: Finish Documentation of the code
TODO: look for optimisation/edgecases